Fixed integer overflow error
It turns out that having ~500k atoms brings about more challenged than
are immediately obvious. I literally ran out of integers to describe the
number of bytes in my dcd file! changing int->long for most places in
the dcdReader seems to alleviate the problem, but a more thorough
accounting of where overflows may occur is needed.

+++ src/reader/dcdReader.hpp
@@ -22,11 +22,11 @@ class dcdReader: public Reader  {
     std::vector<float> box;
 
     float dt;
-    int headerBytes; // bytes
-    int frameBytes; // bytes
-    int frameStart;
-    int frameEnd;
-    int frameSkip;
+    long headerBytes; // bytes
+    long frameBytes; // bytes
+    long frameStart;
+    long frameEnd;
+    long frameSkip;
     char remarks[80];
     char timestamp[80];
